refactoring pagination to use hook
